This is a guessing number games in which you've to guess numbers until you find the right one.
This program was my first game ever on KA and an HTML one this is my second most voted game on KA and it deserves it(not like my most voted program). First I started off with the game generating a random number and checking if the numbers guessed were larger, lesser or the sam(in that case you won). Then I added some good CSS it makes the game look very stylish. :P