A recreation of Pong on Khan Academy
This is a recreation of the Pong game and I'm sure will get more famous than my other programs(this currently has 1 vote due to the fact that I haven't released it). This was one hard game to make(especially in HTML) first I started with the ball movement physics and the player that you are playing as. Then I started to make the AI that was the most hard and fun part to make. I'm sure that this will turn out to be my best program ever on Khan Academy.